Fix Vercel build: avoid metadata route collision by renaming client icon component and updating imports. Prevents server from trying to invoke a Client Component during prerender.
